add discussion button
